Restructure custom class files Implement login integration with backend Add config to admin dashboard
