+++ templates/login.html
@@ -2,10 +2,19 @@
     <div class="row justify-content-center">
         <form class="form-signin" style="width: 400px;" method="post" action="{{ url_for('user_login') }}">
             <h1 class="h3 mb-3 font-weight-normal">User Login</h1>
-            <hr/>
-            <input type="text" id="userID" name="userID" class="form-control" placeholder="User ID" required autofocus>
+            <hr />
+            <input type="text" id="username" name="username" class="form-control" placeholder="User ID" required
+                autofocus>
             <br>
-            <input type="password" name="inputPassword" class="form-control" placeholder="Password" required>
+            <input type="password" name="password" class="form-control" placeholder="Password" required>
+            <div id="registration-div" style="display: none;">
+                <br>
+                <input type="text" id="email" name="email" class="form-control" placeholder="Email Address">
+                <br>
+                <input type="text" id="first_name" name="first_name" class="form-control" placeholder="First Name">
+                <br>
+                <input type="text" id="last_name" name="last_name" class="form-control" placeholder="Last Name">
+            </div>
             <br>
             <div class="checkbox mb-3">
                 <label>
@@ -13,11 +22,52 @@
                 </label>
             </div>
             <div class="btn" style="background-color: var(--color-accent); padding: 0; margin: 0;">
-                <button class="btn btn-outline-dark" style="color: white;" type="submit" name="signin">Sign in</button>
+                <button id="sigbtn" class="btn btn-outline-dark" style="color: white;" type="submit"
+                    name="signin">Sign in</button>
             </div>
             <div class="btn" style="background-color: var(--color-accent); padding: 0; margin: 0;">
-                <button class="btn btn-outline-dark" style="color: white;" type="submit" name="register">Register</button>
+                <button id="regbtn" class="btn btn-outline-dark" style="color: white;" onclick="registration()"
+                    name="register">Register</button>
             </div>
         </form>
     </div>
 </div>
+<script type="text/javascript">
+    var registerFields = ["email", "first_name", "last_name"];
+    var sigbtn = document.getElementById("sigbtn");
+    var regbtn = document.getElementById("regbtn");
+    var regdiv = document.getElementById("registration-div");
+
+    // Reflow form with additional elements for registration
+    function registration() {
+        sigbtn.innerHTML = "Cancel";
+        sigbtn.removeAttribute("type");
+        regbtn.setAttribute("type", "submit");
+
+        registerFields.forEach(field => {
+            var obj = document.getElementById(field);
+            obj.setAttribute("required", "");
+        });
+
+        regdiv.style = "";
+        sigbtn.setAttribute("onclick", "cancelRegistration()");
+        regbtn.removeAttribute("onclick");
+    }
+
+    // Restore form to original layout
+    function cancelRegistration() {
+        regbtn.removeAttribute("type");
+        sigbtn.setAttribute("type", "submit");
+        sigbtn.innerHTML = "Sign In";
+
+        regdiv.style = "display: none;";
+
+        registerFields.forEach(field => {
+            var obj = document.getElementById(field);
+            obj.removeAttribute("required");
+        });
+
+        regbtn.setAttribute("onclick", "registration()");
+        sigbtn.removeAttribute("onclick");
+    }
+</script>
